Move routes, add exception for good.third-party.site
